refactor: update import paths to include file extensions and rename components for consistency
