+++ src/api/socket/useWaitingSocket.ts
@@ -8,11 +8,8 @@ import type {
 } from '@/api/types';
 
 import { useSocket } from './useSocket';
-import { useWaitingStore } from '../store/useWaitingStateStore';
 
 export const useWaitingSocket = () => {
-  const { setUsers, setInvitation, setHostId } = useWaitingStore();
-
   const { socket, connect, disconnect } = useSocket({
     path: 'waiting',
     handshake: '/ws/main-game',
@@ -28,38 +25,27 @@ export const useWaitingSocket = () => {
     if (!socket) return;
 
     const handleWaitingRoomUpdate = (data: WaitingRoomUpdatePayload) => {
-      setUsers(data.users);
+      toast.success(`대기실 업데이트: ${data.users.length}명 참여 중`);
     };
 
     const handleTournamentCreated = (data: TournamentCreatedPayload) => {
       toast.success(`토너먼트가 시작되었습니다. (ID: ${data.tournamentId}, 모드: ${data.mode})`);
-      // Todo: 페이지 이동 추가
-
-      useWaitingStore.getState().clearRoom();
-      useWaitingStore.getState().clearInvitation();
     };
 
     const handleTournamentInvited = (data: CustomInvitedPayload) => {
-      setInvitation(data);
-      setHostId(data.hostId);
-
-      const host = useWaitingStore.getState().users.find((u) => u.id === data.hostId);
-      const nickname = host?.nickname ?? `ID ${data.hostId}`;
+      const nickname = data.hostId ?? `ID ${data.hostId}`;
       toast.info(`${nickname}님이 초대했습니다. 수락하시겠습니까?`, {
         action: {
           label: 'Yes',
           onClick: () => {
             if (socket) {
               socket.emit('custom-accept', { roomId: data.roomId });
-              // TODO: 페이지 이동 추가
             }
           },
         },
         cancel: {
           label: 'No',
-          onClick: () => {
-            useWaitingStore.getState().clearInvitation();
-          },
+          onClick: () => {},
         },
       });
     };
@@ -73,5 +59,5 @@ export const useWaitingSocket = () => {
       socket.off('tournament-created', handleTournamentCreated);
       socket.off('custom-invite', handleTournamentInvited);
     };
-  }, [socket, setUsers, setInvitation, setHostId]);
+  }, [socket]);
 };
