+++ src/constants/routes.ts
@@ -5,10 +5,10 @@ export const PATH = {
   SIGNIN_EMAIL: '/signin/email',
   SIGNUP: '/signup',
   HISTORY: '/history',
-  GAME_SELECT: '/game/select',
-  GAME_AUTO_MATCHING: '/game/auto/matching',
-  GAME_CUSTOM_MATCHING: '/game/custom/matching',
-  TOURNAMENT: '/game/tournament',
+  GAME_SELECT: '/game-select',
+  GAME_AUTO_MATCHING: '/game-select/auto-matching',
+  GAME_CUSTOM_MATCHING: '/game-select/custom-matching',
+  TOURNAMENT: '/game-select/tournament',
 
   FRIEND: '/friend',
   FRIEND_CHATROOM: '/friend/chatroom',
