+++ src/pages/GameSelectPage/GameMatchingPage/AutoGameMatchingPage/Auto1vs1/styles.css.ts
