+++ src/pages/GameSelectPage/GameMatchingPage/AutoGameMatchingPage/AutoTournament/styles.css.ts
