+++ src/pages/GameSelectPage/GameMatchingPage/AutoGameMatchingPage/index.tsx
