+++ src/pages/GameSelectPage/GameMatchingPage/components/invite-friend-dialog/styles.css.ts
