+++ src/pages/GameSelectPage/GameMatchingPage/components/user-card/styles.css.ts
