+++ src/pages/GameSelectPage/TournamentPage/components/Chat/styles.css.ts
