+++ src/pages/GameSelectPage/TournamentPage/components/MatchNode/styles.css.ts
