+++ src/pages/GameSelectPage/TournamentPage/components/PlayerCard/styles.css.ts
