+++ src/pages/GameSelectPage/TournamentPage/styles.css.ts
