+++ src/pages/GameSelectPage/styles.css.ts
