+++ src/pages/index.ts
@@ -6,7 +6,7 @@ export * from './SignInPage';
 export * from './SignUpPage';
 export * from './HistoryPage';
 export * from './ProfilePage';
-export * from './GamePage/TournamentPage';
-export * from './GamePage/GameSelectPage';
-export * from './GamePage/GameMatchingPage/AutoGameMatchingPage';
-export * from './GamePage/GameMatchingPage/CustomGameMatchingPage';
+export * from './GameSelectPage/TournamentPage';
+export * from './GameSelectPage';
+export * from '@/pages/GameSelectPage/GameMatchingPage/AutoGameMatchingPage';
+export * from '@/pages/GameSelectPage/GameMatchingPage/CustomGameMatchingPage';
